+++ vehicle_controller/config/final_01_gps_waypoint.yaml
@@ -1,0 +1,14 @@
+vehicle_controller: # Name of node
+  ros__parameters: # ROS parameters
+    vmax: 3.5
+    angle: 10.0
+    # WP: [latitude, longitude, relative altitude]
+    # 37.465194. 126.950070
+    gps_WP1: [37.2463532, 126.6673070, 60.0] #2
+    gps_WP2: [37.2463532, 126.6693358, 60.0] #3
+    gps_WP3: [37.2479139, 126.6704631, 60.0] #4
+    gps_WP4: [37.2494746, 126.6693359, 60.0] #5
+    gps_WP5: [37.2494746, 126.6670817, 60.0] #6
+    gps_WP6: [37.2479139, 126.6659545, 60.0] #7
+    gps_WP7: [37.2471336, 126.6665181, 60.0] #8
+    gps_WP8: [37.2463532, 126.6670817, 60.0] #home
